requirements.txt
asgiref==3.6.0
beautifulsoup4==4.11.2
Django==4.1.5
django-autoslug==1.9.8
django-bootstrap4==22.3
django-filter==22.1
djangorestframework==3.14.0
Markdown==3.4.1
Pillow==9.4.0
pytz==2022.7.1
soupsieve==2.3.2.post1
sqlparse==0.4.3
tzdata==2022.7
